refactor: メッセージ定数をドメイン層に移動し、レイヤリングを適正化
