+++ src/shared/ipc.ts
@@ -1,5 +1,4 @@
 import type { LogHandler } from '@domain/common/log';
-
 import type { FlacTrack, Picture, ScanResult, TagResult } from '@domain/flac/types';
 import type { Platform } from './platform';
 import type { Unsubscribe } from './types';
